+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support to generate code for `INSERT` queries in `bobgen-mysql`.
 - Added support to generate code for `INSERT`, `UPDATE` and `DELETE` queries in `bobgen-sqlite`.
 - Added `LIMIT` and `OFFSET` to the SQLite Update and Delete queries.
 - Added `IndexedBy` and `NotIndexed` mods to the SQLite Delete queries.